Server/pyproject.toml
[tool.poetry]
name = "ai_academy"
version = "0.0.1"
description = "AI Academy"
authors = ["Sudeep"]
packages = [{ include = "app"}]

[tool.poetry.dependencies]
python = ">=3.12"
python-dotenv = "^1.0.0"
uvicorn = "^0.25.0"
pydantic = "^2.5.0"
sqlalchemy = "^2.0.19"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
email-validator = "^2.1.0"
fastapi = "^0.115.14"
jinja2 = "^3.1.6"
firebase-admin = "^7.1.0"
